Remove poorly specified "class locking"
What does it mean to lock a class ? What does it imply for its inherited classes ? How to implement it ? Object.seal() is not enough as long as a whole instance state is handled by a single object. In this case, it would be impossible to simply manage locking on a per base-class basis. Maybe using a cascade of state holders objects, linked by a __proto__ link ...